Aslam | User expene/ income init.

+++ src/main/java/com/finadv/service/UserService.java
@@ -4,7 +4,7 @@ import com.finadv.entities.User;
 
 public interface UserService {
 
-	User getUserById(int id);
+	User getUserById(Long id);
 	
 	void createUser(User user);
